Refactor vendor and address data models, and update field types for consistency
- Changed `lat` and `lng` fields from `String` or `Double` to `BigDecimal` across `Vendor`, `Address`, and related DTOs (`VendorDTO`, `AddressDTO`, `CartResponseDTO`).
- Updated `deliveryTime` from `String` to `Integer` in `Vendor` and related DTOs.
- Standardized `vendorId` types to `Integer` across repositories, services, and controllers (e.g., `BannerRepository`, `StockRepository`, `VendorController`).
- Updated `Medicine` entity mapping, including renaming the `id` column to `vendor_medicine_id`, mapping `manufacturer` to the `manufacture` column, and marking several fields (`description`, `commonSideEffect`, etc.) as `@Transient`.
- Commented out vendor category-based logic and top-vendor calculation methods in `MainServiceImpl` and `PreOrderServiceImpl`.
- Commented out `findVendorMedicineSalesByCategory` in `OrderItemRepository`.

+++ src/main/java/com/trego/api/VendorController.java
@@ -23,7 +23,7 @@ public class VendorController {
 
 
     @GetMapping("/vendors/{id}")
-    public VendorDTO retrieveVendorById(@PathVariable Long id ,
+    public VendorDTO retrieveVendorById(@PathVariable Integer id ,
                                         @RequestParam(required = false, defaultValue = "") String searchText ,
                                         @RequestParam(defaultValue = "0") int page,
                                         @RequestParam(defaultValue = "10") int size) {
