+++ src/main/java/com/trego/service/IVendorService.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface IVendorService {
     List<VendorDTO> findVendorsByType(String type);
 
-    VendorDTO getVendorByIdOrMedicine(Long id, String searchText , int page, int size);
+    VendorDTO getVendorByIdOrMedicine(Integer id, String searchText , int page, int size);
 }
